+++ admin/src/Statistics/Graphs/ProductGraph.tsx
@@ -5,10 +5,10 @@ import DatePeriodInput, {
     commonPeriodTemplates,
 } from "Components/DatePeriodInput";
 import { RequestParams } from "gateway";
+import { CacheGetter, DataCache, useDataCache } from "Hooks/useDataCache";
+import { usePeriod } from "Hooks/usePeriod";
 import React, { useCallback, useEffect, useMemo } from "react";
 import Select from "react-select";
-import { CacheGetter, DataCache, useDataCache } from "State/useDataCache";
-import { usePeriod } from "State/usePeriod";
 import _ from "underscore";
 import DatePeriod from "../../Models/DatePeriod";
 import { copyArray, sum } from "../../utils";
