+++ admin/src/Statistics/MostActiveMembers.tsx
@@ -1,10 +1,10 @@
 import DatePeriodInput, {
     commonPeriodTemplates,
 } from "Components/DatePeriodInput";
+import { useJson } from "Hooks/useJson";
+import { usePeriod } from "Hooks/usePeriod";
 import React from "react";
 import { Link } from "react-router-dom";
-import { useJson } from "State/useJson";
-import { usePeriod } from "State/usePeriod";
 
 type MemberScore = {
     member_id: number;
